+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support async pipe modules (see [issue 112](https://github.com/KNowledgeOnWebScale/walder/issues/112))
+- Support minimum/maximum of parameters (see [issue 24](https://github.com/KNowledgeOnWebScale/walder/issues/24))
 
 ### Changed
 - Switch to Github actions for CI  (see [issue 102](https://github.com/KNowledgeOnWebScale/walder/issues/102))